Added sticky messages to MAX 7219

+++ lib/CLBCore/src/servoproc.h
@@ -6,7 +6,7 @@
 
 struct ServoSettings {
 	int ServoOutputPin;
-	double ServoInitialPosition;
+	float ServoInitialPosition;
 	bool ServoEnabled;
 };
